refactor(pageindex): consolidate domain helpers and tests
