+++ tests/services/rag/test_pageindex_pipeline.py
@@ -10,9 +10,14 @@ import asyncio
 import json
 from pathlib import Path
 
+import pytest
+
 from deeptutor.services.rag.factory import get_pipeline, normalize_provider_name
 from deeptutor.services.rag.index_versioning import resolve_storage_dir_for_read
+from deeptutor.services.rag.pipelines.pageindex import client as client_mod
 from deeptutor.services.rag.pipelines.pageindex import storage
+from deeptutor.services.rag.pipelines.pageindex.client import PageIndexClient
+from deeptutor.services.rag.pipelines.pageindex.config import PageIndexConfig
 from deeptutor.services.rag.pipelines.pageindex.pipeline import (
     PageIndexPipeline,
     is_supported_file,
@@ -215,3 +220,53 @@ def test_ragservice_resolves_provider_from_metadata(tmp_path) -> None:
     # Explicit override wins over metadata.
     svc_override = RAGService(kb_base_dir=str(tmp_path), provider="llamaindex")
     assert svc_override._resolve_provider("kbx") == "llamaindex"
+
+
+def test_ragservice_guard_prevents_pageindex_pipeline(tmp_path, monkeypatch) -> None:
+    from deeptutor.services.rag.service import RAGService
+
+    service = RAGService(kb_base_dir=str(tmp_path), provider="pageindex-oss")
+    monkeypatch.setattr(
+        service,
+        "_get_pipeline",
+        lambda *_args, **_kwargs: pytest.fail("PageIndex reached a traditional RAG pipeline"),
+    )
+
+    result = asyncio.run(service.search(query="ground this", kb_name="page-kb"))
+
+    assert result["error_type"] == "reasoning_as_retrieval_required"
+
+
+def test_cloud_sdk_client_is_reused_until_api_key_changes(monkeypatch) -> None:
+    created: list[str] = []
+
+    class _CloudClient:
+        def __init__(self, api_key: str) -> None:
+            created.append(api_key)
+
+    monkeypatch.setattr(client_mod, "_sdk_types", lambda: (_CloudClient, object))
+    client_mod._cloud_sdk_client.cache_clear()
+    try:
+        first = PageIndexClient.cloud(PageIndexConfig("key-a")).sdk_client
+        second = PageIndexClient.cloud(PageIndexConfig("key-a")).sdk_client
+        rotated = PageIndexClient.cloud(PageIndexConfig("key-b")).sdk_client
+    finally:
+        client_mod._cloud_sdk_client.cache_clear()
+
+    assert first is second
+    assert rotated is not first
+    assert created == ["key-a", "key-b"]
+
+
+def test_submit_document_uses_sdk_wait() -> None:
+    calls: list[tuple[str, str | None, bool]] = []
+
+    class _SDKClient:
+        def submit_document(self, path: str, *, mode: str | None, wait: bool):
+            calls.append((path, mode, wait))
+            return {"doc_id": "doc-1"}
+
+    doc_id = asyncio.run(PageIndexClient(_SDKClient()).submit_document("a.pdf", mode="flash"))
+
+    assert doc_id == "doc-1"
+    assert calls == [("a.pdf", "flash", True)]
